backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
pillow==11.0.0
google-generativeai==0.8.3
aiohttp==3.10.10
